pyproject.toml
[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "test-analyzer"
version = "0.1.0"
description = "Run tests and use LangChain Deep Agents to analyze CI failures"
readme = "README.md"
requires-python = ">=3.14,<3.15"
dependencies = [
    "click>=8.1.7",
    "deepagents>=0.2.6",
    "langchain-openai>=1.0.2",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
]
build = [
    "pyinstaller>=6.14.0",
]

[project.scripts]
test-analyzer = "test_analyzer.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
